"return IPint.intValue()" to be java 1.4.2
git-svn-id: https://svn.ninux.org/svn/ninuxdeveloping/OBAMPxP/trunk@66 ef3c9aa7-e928-0410-9a28-8491e62b674b

+++ src/Stato.java
@@ -208,7 +208,7 @@ public class Stato /*implements Serializable*/ {
         
         String IPString = i1.toString()+i2.toString()+i3.toString()+i4.toString();
         Integer IPint = new Integer (IPString);
-    	return IPint;
+    	return IPint.intValue();
     }
